+++ gd-smartfarm/src/components/About.jsx
@@ -1,6 +1,0 @@
-// @flow
-import React from 'react';
-
-const About = () => <div>ARPA-E Smartfarm</div>;
-
-export default About;