+++ gd-glm/CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+
+## [3.6.0] - 2020-12-14
+
+### Changed
+- Removed dependencies on gd-glm__old and updated header component
+[GLM-133](https://opensource.ncsa.illinois.edu/jira/browse/GLM-133)
+- Removed analysis page
+[GLM-136](https://opensource.ncsa.illinois.edu/jira/browse/GLM-136)
+
 ## [3.5.0] - 2020-10-01
 ### Changed
 - Decrease the zoom level when a source is selected