+++ gd-gltg/CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [3.6.0] - 2020-12-14
+
+### Added
+- Added district layers
+  [GLGVO-710] (https://opensource.ncsa.illinois.edu/jira/browse/GLGVO-710)
+
+
 ## [3.5.0] - 2020-10-01
 
 ### Changed